Add LunaSec to Key-value Data Model
I wasn't sure if this should go into the Security section or the Key-value Data Model section. Technically, it's a key-value store with security as a focus. None of the other tools in the "security" section seemed very relevant.

If you'd prefer I move it, just let me know. Thank you!